requirements.txt
google-api-core==1.2.0
google-api-python-client==1.7.3
google-cloud-storage==1.10.0
python-telegram-bot==10.1.0
oauth2client==4.1.2
google-cloud==0.32.0
image==1.5.24
